Plugin system updates.   - Refactor.   - Unregistering an existing plugin now works.
